Update structure to match template
