--- requirements.txt
@@ -1,2 +1,0 @@
-mock==4.0.2
-pytest==6.0.1
