+++ setup.cfg
@@ -14,7 +14,6 @@ classifiers =
     License :: OSI Approved :: GNU General Public License (GPL)
     Operating System :: OS Independent
 
-
 [options]
 package_dir = 
     = .
@@ -30,4 +29,4 @@ where = .
 
 [options.entry_points]
 console_scripts =
-    validate = validate.validate_io:validate_main
+    validate = validate.__main__:main
